move throttle enhancer in separated file
fix disabled arrows style
