+++ web/client/reducers/__tests__/geostory-test.js
@@ -123,12 +123,6 @@ describe('geostory reducer', () => {
             expect(sectionAtIndexSelectorCreator(0)(STATE).contents[0].newProp).toBe("PROP");
         });
     });
-    it('geostory updateCurrentPage', () => {
-        const action = updateCurrentPage({sectionId: "ID"});
-        const state = geostory( undefined, action);
-        const currentPage = currentPageSelector({geostory: state});
-        expect(currentPage.sectionId).toBe("ID");
-    });
     it('ADD_RESOURCE', () => {
         expect(
             resourcesSelector({ geostory: geostory({}, addResource("id", "image", {}))})
@@ -208,43 +202,51 @@ describe('geostory reducer', () => {
             })
         ).toBe(SAMPLE_RESOURCE);
     });
-    it('updateCurrentPage', () => {
-        expect(
-            currentPageSelector({
-                geostory: geostory({
-                    currentPage: {
-                        columns: {}
-                    },
-                    currentStory: {
-                        sections: [{
-                            id: "section_1",
-                            contents: [{
-                                id: "column_id_1"
+    describe('updateCurrentPage tests', () => {
+        it('geostory updateCurrentPage, with sectionId', () => {
+            const action = updateCurrentPage({sectionId: "ID"});
+            const state = geostory( undefined, action);
+            const currentPage = currentPageSelector({geostory: state});
+            expect(currentPage.sectionId).toBe("ID");
+        });
+        it('updateCurrentPage, updating columns, with columnId', () => {
+            expect(
+                currentPageSelector({
+                    geostory: geostory({
+                        currentPage: {
+                            columns: {}
+                        },
+                        currentStory: {
+                            sections: [{
+                                id: "section_1",
+                                contents: [{
+                                    id: "column_id_1"
+                                }]
                             }]
-                        }]
-                    }
-                }, updateCurrentPage({columnId: "column_id_1"}))
-            }).columns
-        ).toEqual({ section_1: 'column_id_1' });
-    });
-    it('updateCurrentPage', () => {
-        expect(
-            currentPageSelector({
-                geostory: geostory({
-                    currentPage: {
-                        columns: { section_1: 'column_id_2'}
-                    },
-                    currentStory: {
-                        sections: [{
-                            id: "section_1",
-                            contents: [{
-                                id: "column_id_2"
+                        }
+                    }, updateCurrentPage({columnId: "column_id_1"}))
+                }).columns
+            ).toEqual({ section_1: 'column_id_1' });
+        });
+        it('updateCurrentPage, not updating columns, with columnId', () => {
+            expect(
+                currentPageSelector({
+                    geostory: geostory({
+                        currentPage: {
+                            columns: { section_1: 'column_id_2'}
+                        },
+                        currentStory: {
+                            sections: [{
+                                id: "section_1",
+                                contents: [{
+                                    id: "column_id_2"
+                                }]
                             }]
-                        }]
-                    }
-                }, updateCurrentPage({columnId: "column_id_1"}))
-            }).columns
-        ).toEqual({ section_1: 'column_id_2' });
+                        }
+                    }, updateCurrentPage({columnId: "column_id_1"}))
+                }).columns
+            ).toEqual({ section_1: 'column_id_2' });
+        });
     });
     describe('setControl', () => {
         Object.keys(Controls).forEach(k => {
