+++ web/client/reducers/geostory.js
@@ -19,7 +19,6 @@ import {
     TOGGLE_CARD_PREVIEW,
     UPDATE,
     UPDATE_CURRENT_PAGE,
-    UPDATE_CURRENT_COLUMN,
     REMOVE,
     SELECT_CARD,
     SET_CONTROL,
