+++ web/client/themes/default/less/navigation.less
@@ -24,6 +24,9 @@
         background-color: #ffffff;
         border-bottom: 1px solid #ddd;
     }
+    .scroll-menu-arrow.scroll-menu-arrow--disabled {
+        opacity: 0;
+    }
     @media only screen and (min-width: 601px) {
         .ms-horizontal-menu {
             width: 30%;
